Se realiza ajuste para Factura y sus ítems
Por necesidad de guardar los avances se sube cambios que resuelve el traer de Lectura, Tarifa con todos sus objetos

+++ EdeUNLa/src/negocio/LecturaABM.java
@@ -24,7 +24,7 @@ private static LecturaABM instancia = null; // Patrón Singleton
 		return instancia;
 	}
 
-	public Lectura traer(int idCliente) {
+	public Lectura traer(long idCliente) {
 		return LecturaDao.getInstance().traerLectura(idCliente);
 	}
